miscellanous processed meat

+++ lib/bitcoinj-read-only/src/com/google/bitcoin/core/Indexable.java
@@ -10,7 +10,7 @@ public interface Indexable {
         }
 	};
 	
-	public Long index();
+	public Integer index();
